move hdfs validation tool to integration/validation
