+++ integration/tools/pom.xml
@@ -31,6 +31,7 @@
 
   <modules>
     <module>hms</module>
+    <module>validation</module>
   </modules>
 
   <dependencies>
