+++ shell/pom.xml
@@ -89,6 +89,12 @@
       <artifactId>alluxio-job-client</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.alluxio</groupId>
+      <artifactId>alluxio-integration-tools-validation</artifactId>
+      <version>${project.version}</version>
+      <scope>compile</scope>
+    </dependency>
 
       <!-- Test dependencies -->
     <dependency>
